Handle date bug in Visa table from 2019
This report had an "A" in front of one entry's date.

On the second line where the date is repeated, the "A" causes the date to break onto another line. Track the change in y to continue distinguishing line-breaks from row-breaks when there is more than one line-break in the date column.